Make anaglyph pass and parallax barrier pass extensions ofStereoCompositePassNode class to reuse shared functionality
